Consolidate the JS-only tests and TS tests
